+++ DataLayer/DomainModel/tbTelegramUsers.cs
@@ -32,7 +32,7 @@ namespace DataLayer.DomainModel
         public string Tel_FirstName { get; set; }
         public string Tel_LastName { get; set; }
         public string Tel_RobotID { get; set; }
-        public Nullable<int> Tel_Wallet { get; set; }
+        public Nullable<double> Tel_Wallet { get; set; }
         public Nullable<int> Tel_Parent_ID { get; set; }
         public Nullable<int> Tel_Monthes { get; set; }
         public Nullable<int> Tel_Traffic { get; set; }
